Moved evset _repr_ methods to display_utils
